requirements.txt
absl-py==1.3.0
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
fake-useragent==1.1.1
filelock==3.9.0
Flask==2.1.2
Flask-Cors==3.0.10
flatbuffers==23.1.4
gast==0.4.0
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
h5py==3.7.0
httplib2==0.21.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.11.0
Levenshtein==0.20.9
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
parse==1.19.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==8.2.2
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-Levenshtein==0.20.9
pytz==2022.7
PyYAML==6.0
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.1
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.7.3
sentencepiece==0.1.97
six==1.16.0
snscrape==0.3.4
soupsieve==2.3.2.post1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow
tensorflow-estimator==2.11.0
tensorflow-intel
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
thefuzz==0.19.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
w3lib==2.1.1
websockets==10.4
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.11.0
